requirements.txt
playwright==1.47.0
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2
tenacity==8.5.0
pandas==2.2.2
numpy==2.0.1
beautifulsoup4==4.12.3
lxml==5.2.2
requests==2.32.3
python-dateutil==2.9.0.post0
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic-settings==2.4.0
fastapi==0.115.6
uvicorn==0.30.6
streamlit==1.37.1
plotly==5.23.0
python-dotenv==1.0.1
tenacity==8.5.0
rapidfuzz==3.9.7
apscheduler==3.10.4
loguru==0.7.2
selenium==4.15.2
webdriver-manager==4.0.1